+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -52,6 +52,7 @@ set(SOURCES
     CSS/Parser/Token.cpp
     CSS/Parser/Tokenizer.cpp
     CSS/Percentage.cpp
+    CSS/Position.cpp
     CSS/PreferredColorScheme.cpp
     CSS/Ratio.cpp
     CSS/Resolution.cpp
