+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -36,45 +36,6 @@
 
 namespace Web::CSS {
 
-// FIXME: Named PositionValue to avoid conflicts with enums, but this represents a <position>
-struct PositionValue {
-    enum class HorizontalPreset {
-        Left,
-        Center,
-        Right
-    };
-
-    enum class VerticalPreset {
-        Top,
-        Center,
-        Bottom
-    };
-
-    enum class HorizontalEdge {
-        Left,
-        Right
-    };
-
-    enum class VerticalEdge {
-        Top,
-        Bottom
-    };
-
-    static PositionValue center()
-    {
-        return PositionValue { HorizontalPreset::Center, VerticalPreset::Center };
-    }
-
-    Variant<HorizontalPreset, LengthPercentage> horizontal_position { HorizontalPreset::Left };
-    Variant<VerticalPreset, LengthPercentage> vertical_position { VerticalPreset::Top };
-    HorizontalEdge x_relative_to { HorizontalEdge::Left };
-    VerticalEdge y_relative_to { VerticalEdge::Top };
-
-    CSSPixelPoint resolved(Layout::Node const& node, CSSPixelRect const& rect) const;
-    ErrorOr<void> serialize(StringBuilder&) const;
-    bool operator==(PositionValue const&) const = default;
-};
-
 // FIXME: Find a better place for this helper.
 inline Gfx::Painter::ScalingMode to_gfx_scaling_mode(CSS::ImageRendering css_value)
 {
