+++ Userland/Libraries/LibWeb/CSS/StyleValues/RadialGradientStyleValue.h
@@ -11,6 +11,7 @@
 
 #include <AK/Vector.h>
 #include <LibWeb/CSS/Enums.h>
+#include <LibWeb/CSS/Position.h>
 #include <LibWeb/CSS/StyleValues/AbstractImageStyleValue.h>
 #include <LibWeb/Painting/GradientPainting.h>
